+++ apps/backend/src/app/analytics/sales-analytics/general/repository/pipes/sales-wrapper.pipe.ts
@@ -1,0 +1,49 @@
+import { CandleInterval } from '@common/Enums';
+import { SalesSummery } from '@common/Interfaces';
+import { PipelineStage } from 'mongoose';
+import { salesCandlesPipe } from './sales-candles.pipe';
+import { salesSummaryPipe } from './sales-summery.pipe';
+
+export function emptySummery(): SalesSummery {
+    return {
+        TotalOrdersCount: 0,
+        TotalItemsSold: 0,
+        TotalUniqueProductsSold: 0,
+        TotalUniqueCustomersCount: 0,
+        TotalNewCustomersCount: 0,
+        TotalGrossRevenue: 0,
+    };
+}
+
+//Builds the full  pipeline for general metrics: using the summery and candles pipes
+export function salesGeneralPipe(start: Date, end: Date, interval: CandleInterval, timezone?: string): PipelineStage[] {
+    const summaryStages = salesSummaryPipe(start, end) as PipelineStage.FacetPipelineStage[];
+    const candlesStages = salesCandlesPipe(
+        start,
+        end,
+        interval as CandleInterval,
+        timezone
+    ) as PipelineStage.FacetPipelineStage[];
+
+    return [
+        {
+            $facet: {
+                summary: summaryStages,
+                candles: candlesStages,
+            },
+        },
+        {
+            $project: {
+                summery: {
+                    $ifNull: [
+                        { $arrayElemAt: ['$summary.summery', 0] },
+                        { $literal: emptySummery() }, // inject a constant object
+                    ],
+                },
+                candles: {
+                    $ifNull: [{ $arrayElemAt: ['$candles.candles', 0] }, []],
+                },
+            },
+        },
+    ];
+}
