+++ apps/backend/src/app/analytics/sales-analytics/general/repository/sales-general-repository.interface.ts
@@ -1,7 +1,13 @@
-import { SalesStatsCore, SalesStatsRequestDto } from '../dtos';
+import { CandleInterval } from '@common/Enums';
+import { SalesStatsCore } from '../dtos';
 
 export const SALES_GENERAL_ANALYTICS_REPOSITORY = Symbol('SALES_ANALYTICS_REPOSITORY');
 
 export interface ISalesGeneralAnalyticsRepository {
-    getSalesGeneralMetrics(req: SalesStatsRequestDto): Promise<SalesStatsCore>;
+    getSalesGeneralMetrics(
+        start: Date,
+        end: Date,
+        interval: CandleInterval,
+        timezone?: string
+    ): Promise<SalesStatsCore>;
 }
